+++ src/app/admin/orders/orders.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { SharedModule } from './../../shared/shared.module';
+import { OrderComponent } from './components/order/order.component';
+
+@NgModule({
+  declarations: [OrderComponent],
+  imports: [
+    SharedModule
+  ]
+})
+export class OrdersModule { }