+++ src/app/admin/products/components/product-component/product-component.component.ts
@@ -1,0 +1,27 @@
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { Product } from '../../models/product.model';
+
+@Component({
+  selector: 'app-product-component',
+  templateUrl: './product-component.component.html',
+  styleUrls: ['./product-component.component.css']
+})
+export class ProductComponentComponent {
+
+  @Input() product: Product;
+
+  @Output() buyProduct: EventEmitter<Product> = new EventEmitter<Product>();
+
+  public onEditProduct(): void {
+    this.buyProduct.emit(this.product);
+  }
+
+  public setClasses() {
+    const classes = {
+      available: this.product.isAvailable,
+      notAvailable: !this.product.isAvailable
+    };
+    return classes;
+  }
+
+}