+++ src/app/admin/products/products-routing.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { ProductListComponentComponent, ProductFormComponent } from './components';
+
+const routes: Routes = [
+  {
+    path: 'products',
+    component: ProductListComponentComponent
+  },
+  {
+    path: '',
+    redirectTo: '/products',
+    pathMatch: 'full'
+  },
+  {
+    path: 'edit/:id',
+    component: ProductFormComponent
+  }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes)
+  ],
+  exports: [RouterModule]
+})
+export class ProductsRoutingModule {}