+++ src/app/cart/cart.module.ts
@@ -3,6 +3,7 @@ import { NgModule } from '@angular/core';
 import { CartComponentComponent } from './components/cart-component/cart-component.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 import { SharedModule } from './../shared/shared.module';
+import { CartRoutingModule } from './cart-routing.module';
 
 @NgModule({
   declarations: [
@@ -10,7 +11,8 @@ import { SharedModule } from './../shared/shared.module';
     CartItemComponent
   ],
   imports: [
-    SharedModule
+    SharedModule,
+    CartRoutingModule
   ],
   exports: [
     CartComponentComponent