+++ src/app/layout/components/index.ts
@@ -1,0 +1,4 @@
+export * from './about/about.component';
+export * from './home/home.component';
+export * from './path-not-found/path-not-found.component';
+export * from './login/login.component';