+++ src/app/orders/models/order.model.ts
@@ -1,0 +1,7 @@
+export class Order {
+    constructor(
+        public price: number,
+        public count: number,
+        public created: Date) {
+    }
+}