+++ src/app/orders/orders-routing.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { OrderComponent } from './components/order/order.component';
+
+const routes: Routes = [
+  {
+    path: 'order',
+    component: OrderComponent
+  },
+  {
+    path: '',
+    redirectTo: '/order',
+    pathMatch: 'full'
+  }
+];
+
+@NgModule({
+  imports: [
+    RouterModule.forChild(routes)
+  ],
+  exports: [RouterModule]
+})
+export class OrdersRoutingModule {}