+++ src/app/shared/enums/enums.ts
@@ -4,4 +4,3 @@ enum Category {
 }
 
 export { Category };
-// этот файл надо из корня приложения перенести, например в shared/enums